refactor: use new BaseTheme class and cleanup class registration. move base.php into views folder
