+++ app/ThemeFilters.php
@@ -2,22 +2,19 @@
 
 namespace Vibrato;
 
-use Vibrato\Traits\Booted;
+use Vibrato\Core\BaseTheme;
 
 /**
  * class ThemeFilters
  *
  * @since 1.0.0
  */
-class ThemeFilters
+class ThemeFilters extends BaseTheme
 {
-    use Booted;
-
     public $primary_navs = ['primary_navigation'];
 
     protected function register()
     {
-        add_filter('template_include', ['Vibrato\\ThemeTemplateWrapper', 'wrap'], 109);
         add_filter('body_class', array($this, 'body_class'));
         add_filter('excerpt_more', array($this, 'excerpt_more'));
         add_filter('mce_buttons_2', array($this, 'custom_tinymce_buttons'));
