+++ app/ThemeSetup.php
@@ -2,16 +2,15 @@
 
 namespace Vibrato;
 
-use Vibrato\Traits\Booted;
+use Vibrato\Core\BaseTheme;
 
-class ThemeSetup
+class ThemeSetup extends BaseTheme
 {
-    use Booted;
-
     protected function register()
     {
         add_action('after_setup_theme', array($this, 'setup'));
         add_action('wp_enqueue_scripts', array($this, 'resources'), 100);
+        add_filter('template_include', [ThemeTemplateWrapper::class, 'wrap'], 109);
     }
 
     /**
