+++ resources/views/common/base.php
